+++ tests/src/App/Entity/Country.php
@@ -16,9 +16,9 @@ namespace Rekalogika\Collections\Tests\App\Entity;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
 use Doctrine\ORM\Mapping as ORM;
-use Rekalogika\Collections\Tests\App\Repository\CountryRepository;
+use Rekalogika\Collections\Tests\App\DoctrineRepository\DoctrineCountryRepository;
 
-#[ORM\Entity(repositoryClass: CountryRepository::class)]
+#[ORM\Entity(repositoryClass: DoctrineCountryRepository::class)]
 class Country
 {
     #[ORM\Id]
@@ -34,6 +34,9 @@ class Country
      */
     #[ORM\OneToMany(targetEntity: Citizen::class, mappedBy: 'country', orphanRemoval: true)]
     private Collection $citizens;
+
+    #[ORM\Column(length: 255)]
+    private ?string $code = null;
 
     public function __construct()
     {
@@ -83,6 +86,18 @@ class Country
                 $citizen->setCountry(null);
             }
         }
+
+        return $this;
+    }
+
+    public function getCode(): ?string
+    {
+        return $this->code;
+    }
+
+    public function setCode(string $code): static
+    {
+        $this->code = $code;
 
         return $this;
     }