+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 * test: add skeleton tests using Symfony framework
 * fix(`BasicRepository`): `remove()` should be `removeElement()`
 * feat(`BasicRepository`): add `remove()` method
+* test: add basic repository test
 
 ## 0.3.0